check channel_reserve_satoshis and dust_limit_satoshis
