+++ ln/Makefile
@@ -38,6 +38,7 @@ remduplicates = $(strip $(if $1,$(firstword $1) $(call remduplicates,$(filter-ou
 #sources project
 C_SOURCE_FILES += $(PRJ_PATH)/ln.c
 C_SOURCE_FILES += $(PRJ_PATH)/ln_script.c
+C_SOURCE_FILES += $(PRJ_PATH)/ln_comtx.c
 C_SOURCE_FILES += $(PRJ_PATH)/ln_derkey.c
 C_SOURCE_FILES += $(PRJ_PATH)/ln_msg_establish.c
 C_SOURCE_FILES += $(PRJ_PATH)/ln_msg_close.c
